+++ package.json
@@ -12,6 +12,7 @@
     "mongoose": "^5.12.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "socket.io": "^4.0.1",
     "styled-components": "^5.2.3"
