--- src/App.js
@@ -1,29 +1,0 @@
-const App = () => {
-  return <div>Here's the app
-    <div><h>Logo</h><h>thinkquiry.io</h></div>
-        <div>
-          <h>Thinkquiry!</h>
-          <h>Join Room!</h>
-          <input placeholder='Room'></input>
-          <input placeholder='Name'></input>
-          <button>Enter!</button>
-        </div>
-        <div>
-          <h>Join Existing Room As Admin</h>
-          <input placeholder='Room'></input>
-          <input placeholder='NAME TOO?'></input>
-          <input placeholder='password'></input>
-          <button>Enter</button>
-        </div>
-        <div>
-          <h>Create a New Room</h>
-          <input placeholder='Room Name'></input>
-          <input placeholder='Create Password'></input>
-          <button>Enter</button>
-        <div>
-      </div>
-      </div>
-    </div>;
-};
-
-export default App;
