+++ src/components/Home.jsx
@@ -1,0 +1,36 @@
+import styled from 'styled-components';
+
+const Home = () => {
+  return (
+    <div>
+      Here's the app
+      <div>
+        <h>Logo</h>
+        <h>thinkquiry.io</h>
+      </div>
+      <div>
+        <h>Thinkquiry!</h>
+        <h>Join Room!</h>
+        <input placeholder='Room'></input>
+        <input placeholder='Name'></input>
+        <button>Enter!</button>
+      </div>
+      <div>
+        <h>Join Existing Room As Admin</h>
+        <input placeholder='Room'></input>
+        <input placeholder='NAME TOO?'></input>
+        <input placeholder='password'></input>
+        <button>Enter</button>
+      </div>
+      <div>
+        <h>Create a New Room</h>
+        <input placeholder='Room Name'></input>
+        <input placeholder='Create Password'></input>
+        <button>Enter</button>
+        <div></div>
+      </div>
+    </div>
+  );
+};
+
+export default Home;
